requirements.txt
# Основные библиотеки для работы с Android и ADB
pure-python-adb==0.3.0.dev0    # Управление ADB без внешних зависимостей
#uiautomator2==2.16.23          # Альтернатива для UI автоматизации (опционально)

# Обработка изображений и компьютерное зрение
opencv-python==4.12.0.88        # Распознавание шаблонов и элементов UI
#Pillow==25.0                 # Работа с изображениями
#numpy==2.2.6                  # Поддержка массивов для OpenCV

# OCR для чтения текста (опционально)
pytesseract==0.3.10            # OCR для чтения чисел и таймеров
# paddleocr==2.7.0             # Альтернативный OCR (раскомментировать при необходимости)

# Планирование и управление задачами
schedule==1.2.0                # Планировщик задач
APScheduler==3.10.4            # Продвинутый планировщик (опционально)

# Логирование и мониторинг
loguru==0.7.2                  # Продвинутое логирование
colorama==0.4.6                # Цветной вывод в консоль

# Работа с конфигурациями
PyYAML==6.0.1                  # Парсинг YAML конфигов

# Системный мониторинг
psutil==5.9.6                  # Мониторинг CPU, RAM, процессов

# База данных
# SQLite встроен в Python, дополнительные библиотеки не требуются

# Дополнительные утилиты
python-dotenv==1.0.0           # Загрузка переменных окружения
click==8.1.7                   # CLI интерфейс
tqdm==4.66.1                   # Прогресс-бары

# Веб-интерфейс (опционально, для этапа 8)
# streamlit==1.28.2            # Веб-дашборд
# flask==3.0.0                 # Альтернатива для веб-интерфейса